+++ src/rollem-bot/behaviors/parse-soft.behavior.ts
@@ -1,9 +1,9 @@
 import { RollBehaviorBase } from "./roll-behavior-base";
-import { Parsers } from "@bot/lib/parsers";
 import { Client } from "discord.js";
 import { Logger } from "@bot/logger";
 import { Config } from "@bot/config";
 import { Injectable } from "injection-js";
+import { Parsers } from "@bot/lib/parsers";
 
 /**
  * Parses things without any prefix, unless a prefix is configured via role-name.
@@ -16,7 +16,7 @@ export class ParseSoftBehavior extends RollBehaviorBase {
     client: Client,
     logger: Logger,
   ) { super(parsers, config, client, logger); }
-  
+
   protected register() {
     this.client.on('message', message => {
       // avoid doing insane things
@@ -24,17 +24,17 @@ export class ParseSoftBehavior extends RollBehaviorBase {
       if (message.author == this.client.user) { return; }
       if (this.shouldDefer(message)) { return; }
       if (message.content.startsWith('D')) { return; } // apparently D8 is a common emote.
-    
+
       // honor the prefix
-      let prefix = this.getPrefix(message);
+      const prefix = this.getPrefix(message);
       if (!message.content.startsWith(prefix)) { return; }
-    
+
       // get our actual roll content
       let content = message.content.substring(prefix.length);
       content = content.trim();
-      
-      let requireDice = true;
-      let lines = this.rollMany(message, "No-prefix roll", content, !!prefix, requireDice);
+
+      const requireDice = true;
+      const lines = this.rollMany(message, "No-prefix roll", content, !!prefix, requireDice);
       this.replyAndLog(message, 'No-prefix roll', lines);
     });
   }