+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# **1.15.3**
+# **1.15.4**
 * Fix crit-range display in output.
 * Add ko-fi + patreon callouts to status message.
 * Start work on v2 of the rollem syntax. Nowhere near ready yet!